backend: Make pipeline a top-level component.
And move it to API.
Plus other code moving around.
