+++ packages/backend-api/bin/osmod.js
@@ -35,6 +35,9 @@ yargs
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'users', 'create')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'users', 'get_token')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'tests', 'youtube')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'rescore')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'send_to_scorer')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'generate')))
   .demand(1)
   .usage('Usage: $0')
   .help()
