+++ packages/backend-api/src/pipeline/shim.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {ICommentInstance} from '../../models';
+import { ICommentInstance } from '@conversationai/moderator-backend-core';
 
 export interface IScore {
   score: number;
