+++ packages/backend-api/src/processing/tasks/send_comment_for_scoring.ts
@@ -17,8 +17,9 @@ limitations under the License.
 import {
   Article,
   Comment,
-  sendForScoring,
 } from '@conversationai/moderator-backend-core';
+
+import { sendForScoring } from '../../pipeline';
 import { handler, IQueueHandler } from '../util';
 
 export interface ISendCommentForScoringTaskData {
