+++ packages/backend-api/src/test/integration/api/assignments.spec.ts
@@ -41,7 +41,7 @@ import {
   makeCategory,
   makeComment,
   makeUser,
-} from '../../test_helper';
+} from '../../fixture';
 import {
   app,
 } from './test_helper';
