+++ packages/backend-api/src/test/integration/api/assistant.spec.ts
@@ -30,7 +30,7 @@ import {
   expect,
   makeComment,
   makeUser,
-} from '../../test_helper';
+} from '../../fixture';
 import {
   app,
 } from './test_helper';
