+++ packages/backend-api/src/test/integration/api/authorCounts.spec.ts
@@ -24,7 +24,7 @@ import {
 import {
   expect,
   makeComment,
-} from '../../test_helper';
+} from '../../fixture';
 import {
   app,
 } from './test_helper';
