+++ packages/backend-api/src/test/integration/api/publisher.spec.ts
@@ -25,7 +25,7 @@ import {
   makeArticle,
   makeCategory,
   makeComment,
-} from '../../test_helper';
+} from '../../fixture';
 import {
   app,
 } from './test_helper';
