+++ packages/backend-api/src/test/unit/services/authorCounts.spec.ts
@@ -21,7 +21,7 @@ import {
 import {
   expect,
   makeComment,
-} from '../../test_helper';
+} from '../../fixture';
 
 import {
   getAuthorCounts,
