+++ packages/backend-api/src/test/unit/services/histogramScores.spec.ts
@@ -33,7 +33,7 @@ import {
   makeComment,
   makeCommentSummaryScore,
   makeTag,
-} from '../../test_helper';
+} from '../../fixture';
 
 import {
   getHistogramScoresForAllCategories,
