+++ packages/backend-core/bin/osmod.js
@@ -38,10 +38,7 @@ yargs
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'recalculate_text_sizes')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'calculate_text_size')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'recalculate_top_scores')))
-  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'rescore')))
-  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'send_to_scorer')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'delete')))
-  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'generate')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'flag')))
   .demand(1)
   .usage('Usage: $0')
